+++ objectified-ui/public/WHATS_NEW.md
@@ -113,6 +113,7 @@ We continue to improve the platform based on your feedback with improvements and
   - Adds the ability to map external types to internal types using type mapping during imports
   - Adds default value assignment during imports by setting global defaults during import
   - Adds the ability to override the required field during property mapping at import
+  - Adds/modifies descriptions for property mapping during import
 - OpenAPI Support:
   - Now at 100% support for class and properties definitions
   - Improved import support for paths and securitySchemes